requirements.txt
transformers
Pillow
ml-collections
torch
torchvision
setproctitle
torchinfo
numpy
pyyaml
timm == 0.4.12
mmcv==1.3.8
mmsegmentation==0.14.1